+++ src/main/entity/OneTimeOutflow.java
@@ -3,7 +3,7 @@ package entity;
 import java.time.LocalDate;
 
 public class OneTimeOutflow extends OneTimeTransaction {
-    public OneTimeOutflow(String identification, double amount, LocalDate date, String description) {
+    public OneTimeOutflow(String identification, float amount, LocalDate date, String description) {
         super(identification, amount, date, description);
         if(this.isInflow()) System.out.println("Periodic outflow error, should be inflow!");
     }
